Add files via upload
To complete the assignment, follow the steps below. The `titanic.csv` is attached. Steps 1, 2, and 3 are independent of each other, steps 4 onwards depend on the previous ones. Every step has an example at the end of this text of how it should behave.

Write a function `fill_age` that receives a parameter `value` of type `str`, and returns a `float` if the parameter is not empty, otherwise returns `-1`. (1 point)
Write a function `split_raw_fields` that receives a parameter `line` of type `str`, containing several words separated by commas, and returns a `list` of `str`, where each element corresponds to each of the words of the original string, without any whitespace at the beginning or at the end of the word. (2 points)
Write a function `build_record` that receives two parameters: `fields` (a `tuple` of heterogeneous elements) and `header` (a `list` of `str`) and returns a `dict`, where each key is an element of `header` and each value is an element of `fields`. (2 points)
Write a function `extract_fields` that receives a parameter `line` of type `str` with a specific structure (see example below), uses `fill_age` and `split_raw_fields`, and returns a `tuple` of elements, with the types (int, bool, int, str, str, float, int, int, str, float, str, str). (2 points)
Write a function `read_data` that receives a file handle (as returned by `open`) and returns a `list` of records, each record corresponding to each line of `titanic.csv` (except for the first one). Hint: store the first line as `header`, you will need it to call `build_record` on the rest of the lines of the file. (2 points)
Compute the mean age of the first 10 records, excluding the ones where the age is unknown (where "unknown", given our `fill_age` function, is `-1`). (1 point)